auth_service/requirements.txt
Flask
Flask-CORS
PyJWT
streamlit
requests
pandas
plotly
python-dotenv
gunicorn